requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0
whitenoise==6.6.0
dj-database-url==2.2.0
